- [BE] feat(document): 등기부등본 스키마 수정에 따른 코드 변경 (S14P21C103-266)

See merge request s14-fintech-finance-sub1/S14P21C103!99

+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/api/controller/game/realestate/RealEstateDocumentController.java
@@ -18,8 +18,8 @@ public class RealEstateDocumentController {
 
     @GetMapping("/{propertyId}/documents")
     public ApiResponse<RealEstateDocumentResponse> getDocuments(
-        @PathVariable final Integer sessionId,
-        @PathVariable final Integer propertyId
+        @PathVariable final Long sessionId,
+        @PathVariable final Long propertyId
     ) {
         final RealEstateDocumentResponse response = realEstateDocumentService.getDocument(sessionId, propertyId);
         return ApiResponse.ok(response);
